requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
passlib==1.6.2
Flask-Uploads==0.1.3
celery==3.1.17
sqlalchemy_utils==0.28.2
Flask-Mail==0.9.1
Flask-Images==2.1.2
